requirements.txt
Pillow==10.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
transformers==4.51.2
python-dotenv==1.0.0
torch==2.7.0
tqdm==4.66.1
ftfy==6.1.1
rapidfuzz==3.8.1
surya-ocr==0.14.2
regex==2024.4.28
pdftext~=0.6.2
markdownify==0.13.1
click==8.2.0
markdown2==2.5.2
filetype==1.2.0
scikit-learn==1.6.1
google-genai==1.0.0
anthropic==0.46.0
pre-commit==4.2.0
mammoth
streamlit
streamlit_ace
openai
fastapi
uvicorn
python-multipart